@@ -5,7 +5,8 @@ author:
 homepage:
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+#  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
   flutter: ">=1.20.0"
 
 dependencies:
